Strip redundant changes
