+++ impl/Cargo.toml
@@ -59,7 +59,7 @@ deref_mut = []
 display = ["syn/extra-traits", "dep:unicode-xid", "dep:convert_case"]
 error = ["syn/extra-traits"]
 from = ["syn/extra-traits"]
-from_str = ["dep:convert_case"]
+from_str = []
 index = []
 index_mut = []
 into = ["syn/extra-traits"]
